requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=0.24.2
ta>=0.7.0
plotly>=5.1.0
dash>=2.0.0
python-binance>=1.0.15
python-dotenv>=0.19.0
SQLAlchemy>=1.4.0
aiohttp>=3.8.0
websockets>=10.0
openai>=1.3.0
pyyaml>=6.0.1
matplotlib>=3.7.1
tqdm>=4.65.0
pytest>=7.3.1
jupyter>=1.0.0
requests>=2.28.0
python-dateutil>=2.8.2
pytz>=2023.3
flask>=2.0.1
flask-sqlalchemy>=2.5.1
dash-bootstrap-components>=1.0.0
gunicorn>=20.1.0
pandas-ta==0.3.14b0
ccxt>=3.0.0
websocket-client>=1.2.0
asyncio>=3.4.3
python-telegram-bot>=13.7
redis>=4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
